+++ blocklets/ai-studio/api/src/store/models/project-extra.ts
@@ -1,0 +1,40 @@
+import { CreationOptional, DataTypes, InferAttributes, InferCreationAttributes, Model } from 'sequelize';
+import { Worker } from 'snowflake-uuid';
+
+import { sequelize } from '../sequelize';
+
+const idGenerator = new Worker();
+
+const nextId = () => idGenerator.nextId().toString();
+
+export default class ProjectExtra extends Model<InferAttributes<ProjectExtra>, InferCreationAttributes<ProjectExtra>> {
+  declare id: CreationOptional<string>;
+
+  declare createdAt: CreationOptional<Date>;
+
+  declare updatedAt: CreationOptional<Date>;
+
+  declare npmPackageSecret?: string;
+}
+
+ProjectExtra.init(
+  {
+    id: {
+      type: DataTypes.STRING,
+      primaryKey: true,
+      allowNull: false,
+      defaultValue: nextId,
+    },
+    createdAt: {
+      type: DataTypes.DATE,
+    },
+    updatedAt: {
+      type: DataTypes.DATE,
+    },
+    npmPackageSecret: {
+      type: DataTypes.STRING,
+      allowNull: true,
+    },
+  },
+  { sequelize }
+);