+++ blocklets/ai-runtime/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.4.183 (2024-12-20)
+
+- chore(project-models): ui improvements
+
+## 0.4.182 (2024-12-20)
+
+- Support install a project as npm package
+
 ## 0.4.177 (2024-12-18)
 
 - fix: data not loading after changing resource type