+++ framework/runtime/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @aigne/runtime
+## 0.0.15
+
+- Support install a project as npm package
 
 ## 0.0.14
-
-### Patch Changes
 
 - add aigne-cli to generate stub code